spllit RPC clients into submodules
