+++ justfile
@@ -38,3 +38,18 @@ run-jupyter jupyter_args="":
 
 build runtime:
     docker build -t syft_python_runtime .
+
+
+[group('test')]
+run-rds-integration-tests:
+    uv run --frozen --with "pytest" \
+        pytest syft-rds/tests/integration/crud_test.py
+
+[group('test')]
+run-rds-unit-tests:
+    uv run pytest syft-rds/tests/unit/*_test.py
+
+[group('test')]
+run-tests:
+    just run-rds-unit-tests
+    just run-rds-integration-tests
