+++ syft-rds/src/syft_rds/client/rpc_client.py
@@ -1,44 +1,15 @@
-from typing import TYPE_CHECKING, ClassVar, Generic, TypeVar
-
-from pydantic import BaseModel
+from typing import TYPE_CHECKING
 
 from syft_rds.client.connection import RPCConnection
-from syft_rds.models.base import BaseSchema, BaseSchemaCreate, BaseSchemaUpdate
-from syft_rds.models.models import (
-    Dataset,
-    DatasetCreate,
-    DatasetUpdate,
-    GetAllRequest,
-    GetOneRequest,
-    Job,
-    JobCreate,
-    JobUpdate,
-    Runtime,
-    RuntimeCreate,
-    RuntimeUpdate,
-    UserCode,
-    UserCodeCreate,
-    UserCodeUpdate,
-)
+from syft_rds.client.rpc_clients.base import RPCClientModule
+from syft_rds.client.rpc_clients.jobs import JobRPCClient
+from syft_rds.client.rpc_clients.user_code import UserCodeRPCClient
+from syft_rds.client.rpc_clients.runtime import RuntimeRPCClient
+from syft_rds.client.rpc_clients.dataset import DatasetRPCClient
+
 
 if TYPE_CHECKING:
     from syft_rds.client.rds_client import RDSClientConfig
-
-
-class RPCClientModule:
-    def __init__(self, config: "RDSClientConfig", connection: RPCConnection):
-        self.config = config
-        self.connection = connection
-
-        self.prefix = f"syft://{self.config.host}/api_data/{self.config.app_name}/rpc"
-
-    def _send(self, path: str, body: BaseModel) -> dict:
-        return self.connection.send(
-            f"{self.prefix}/{path}",
-            body,
-            expiry=self.config.rpc_expiry,
-            cache=self.config.rpc_cache,
-        )
 
 
 class RPCClient(RPCClientModule):
@@ -48,40 +19,3 @@ class RPCClient(RPCClientModule):
         self.user_code = UserCodeRPCClient(self.config, self.connection)
         self.runtime = RuntimeRPCClient(self.config, self.connection)
         self.dataset = DatasetRPCClient(self.config, self.connection)
-
-
-T = TypeVar("T", bound=BaseSchema)
-CreateT = TypeVar("CreateT", bound=BaseSchemaCreate)
-UpdateT = TypeVar("UpdateT", bound=BaseSchemaUpdate)
-
-
-class CRUDRPCClient(RPCClientModule, Generic[T, CreateT, UpdateT]):
-    MODULE_NAME: ClassVar[str]
-
-    def create(self, item: CreateT) -> T:
-        return self._send(f"{self.MODULE_NAME}/create", item)
-
-    def get_one(self, request: GetOneRequest) -> T:
-        return self._send(f"{self.MODULE_NAME}/get_one", request)
-
-    def get_all(self, request: GetAllRequest) -> list[T]:
-        return self._send(f"{self.MODULE_NAME}/get_all", request)
-
-    def update(self, item: UpdateT) -> T:
-        return self._send(f"{self.MODULE_NAME}/update", item)
-
-
-class JobRPCClient(CRUDRPCClient[Job, JobCreate, JobUpdate]):
-    MODULE_NAME = "job"
-
-
-class UserCodeRPCClient(CRUDRPCClient[UserCode, UserCodeCreate, UserCodeUpdate]):
-    MODULE_NAME = "user_code"
-
-
-class RuntimeRPCClient(CRUDRPCClient[Runtime, RuntimeCreate, RuntimeUpdate]):
-    MODULE_NAME = "runtime"
-
-
-class DatasetRPCClient(CRUDRPCClient[Dataset, DatasetCreate, DatasetUpdate]):
-    MODULE_NAME = "dataset"
