+++ syft-rds/src/syft_rds/client/rpc_clients/user_code.py
@@ -1,0 +1,6 @@
+from syft_rds.client.rpc_clients.base import CRUDRPCClient
+from syft_rds.models.models import UserCode, UserCodeCreate, UserCodeUpdate
+
+
+class UserCodeRPCClient(CRUDRPCClient[UserCode, UserCodeCreate, UserCodeUpdate]):
+    MODULE_NAME = "user_code"
